Changed read buffer type to double

+++ src/resample.h
@@ -8,8 +8,8 @@
 #include <stdlib.h>
 #include "wav_file.h"
 
-int16_t* upsample(const int16_t* source, size_t len, int factor);
-int16_t* downsample(const int16_t* source, size_t len, int factor);
+double* upsample(const double* source, size_t len, int factor);
+double* downsample(const double* source, size_t len, int factor);
 
 void resample_wave_file(WAV_FILE* wav_file,
                         int upsample_factor,
